requirements.txt
matplotlib==3.10.7
numpy==1.23.5
PyYAML==5.3
PyYAML==6.0.3
scikit_learn==1.7.2
timm==0.5.4
torch==2.5.1+cu118
torchvision==0.20.1+cu118
